SOT fixes, phase 7

+++ apps/common/r01_play_camera.h
@@ -1,6 +1,8 @@
 #ifndef R01_PLAY_CAMERA_H
 #define R01_PLAY_CAMERA_H
 
+#include <stdint.h>
+
 #define R01_PLAY_CAM_DEADZONE_X_DEFAULT 32
 #define R01_PLAY_CAM_DEADZONE_Y_DEFAULT 30
 
@@ -24,11 +26,13 @@
 
 /* Packed dead-zone width/height. Follow snaps edges to viewport-center
  * parity (docs/general/world-scrolling.md). Live box may be 1 px smaller. */
-void r01_play_camera_update(int *cam_x, int *cam_y, int anchor_x, int anchor_y, int player_w, int player_h,
-                            int screen_w, int screen_h, int deadzone_x, int deadzone_y, int axis_lock);
+void r01_play_camera_update(uint16_t *cam_x, uint16_t *cam_y, uint16_t anchor_x, uint16_t anchor_y,
+                            uint8_t player_w, uint8_t player_h, uint8_t screen_w, uint8_t screen_h,
+                            uint8_t deadzone_x, uint8_t deadzone_y, uint8_t axis_lock);
 
 /* Ensure anchor (entity origin) is inside the dead zone; call before first frame after spawn/warp. */
-void r01_play_camera_snap(int *cam_x, int *cam_y, int anchor_x, int anchor_y, int player_w, int player_h,
-                          int screen_w, int screen_h, int deadzone_x, int deadzone_y, int axis_lock);
+void r01_play_camera_snap(uint16_t *cam_x, uint16_t *cam_y, uint16_t anchor_x, uint16_t anchor_y, uint8_t player_w,
+                          uint8_t player_h, uint8_t screen_w, uint8_t screen_h, uint8_t deadzone_x,
+                          uint8_t deadzone_y, uint8_t axis_lock);
 
 #endif
